requirements.txt
torch
torchvision
gradio
pillow
mlflow
boto3
sagemaker